--- a/app/(auth)/login/loginForm.tsx
+++ b/app/(auth)/login/loginForm.tsx
@@ -2,7 +2,6 @@
 
 import { ArrowRight } from "lucide-react";
 import { AnimatePresence, motion } from "motion/react";
-import { useTheme } from "next-themes";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
@@ -28,7 +27,6 @@ export function LoginForm({ className, ...props }: React.ComponentPropsWithoutRe
   const [isLoading, setIsLoading] = useState(false);
   const [dashboardUrl, setDashboardUrl] = useState<string | null>(null);
   const router = useRouter();
-  const { theme, systemTheme } = useTheme();
 
   const startSignInMutation = api.user.startSignIn.useMutation({
     onSuccess: (data) => {
@@ -115,7 +113,7 @@ export function LoginForm({ className, ...props }: React.ComponentPropsWithoutRe
     <div className={cn("flex flex-col gap-6", className)} {...props}>
       <div className="flex flex-col items-center gap-3">
         <Image
-          src={theme === "dark" || systemTheme === "dark" ? "/logo-white.svg" : "/logo.svg"}
+          src="/logo.svg"
           alt="Helper"
           width="110"
           height="32"