--- a/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/tabBar.tsx
+++ b/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/tabBar.tsx
@@ -21,16 +21,8 @@ const tabsSchema = z.object({
 
 type Tab = z.infer<typeof tabsSchema>["tabs"][number];
 
-const generateUUID = () => {
-  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
-    const r = Math.random() * 16 | 0;
-    const v = c == 'x' ? r : (r & 0x3 | 0x8);
-    return v.toString(16);
-  });
-};
-
 const newTab = (url?: string) => {
-  return { id: generateUUID(), title: document.title, url: url ?? location.href };
+  return { id: crypto.randomUUID(), title: document.title, url: url ?? location.href };
 };
 
 const buildFirstTab = () => {