--- a/app/(dashboard)/mailboxes/[mailbox_slug]/clientLayout.tsx
+++ b/app/(dashboard)/mailboxes/[mailbox_slug]/clientLayout.tsx
@@ -1,8 +1,6 @@
 "use client";
 
-import { createContext, useContext, useMemo, useState } from "react";
 import { create } from "zustand";
-import { buildThemeCss, MailboxTheme } from "@/lib/themes";
 
 export const useShowChatWidget = create<{
   showChatWidget: boolean;
@@ -12,29 +10,16 @@ export const useShowChatWidget = create<{
   setShowChatWidget: (showChatWidget) => set({ showChatWidget }),
 }));
 
-const InboxThemeContext = createContext<{
-  theme: MailboxTheme | undefined;
-  setTheme: (theme: MailboxTheme | undefined) => void;
-}>({
-  theme: undefined,
-  setTheme: () => {},
-});
-
-export const useInboxTheme = () => useContext(InboxThemeContext);
 
 export default function InboxClientLayout({
   children,
-  theme: initialTheme,
 }: {
   children: React.ReactNode;
-  theme?: MailboxTheme;
 }) {
-  const [theme, setTheme] = useState<MailboxTheme | undefined>(initialTheme);
   const { showChatWidget } = useShowChatWidget();
-  const themeCss = useMemo(() => buildThemeCss(theme), [theme]);
 
   return (
-    <InboxThemeContext.Provider value={{ theme, setTheme }}>
+    <>
       {/* We show the widget for testing on the chat settings page. Need to improve the SDK to allow destroying the widget so we can move the provider there */}
       {!showChatWidget && (
         <style>
@@ -45,8 +30,7 @@ export default function InboxClientLayout({
           `}
         </style>
       )}
-      <style>{themeCss}</style>
       {children}
-    </InboxThemeContext.Provider>
+    </>
   );
 }