--- a/app/(dashboard)/mailboxes/[mailbox_slug]/layout.tsx
+++ b/app/(dashboard)/mailboxes/[mailbox_slug]/layout.tsx
@@ -16,12 +16,11 @@ export default async function InboxLayout({
 }) {
   try {
     const mailboxSlug = (await params).mailbox_slug;
-    const { preferences } = await api.mailbox.get({ mailboxSlug });
 
     return (
       <HelperProvider host={env.AUTH_URL} mailboxSlug={mailboxSlug} showToggleButton>
         <SidebarProvider>
-          <InboxClientLayout theme={preferences?.theme}>
+          <InboxClientLayout>
             <div className="flex h-svh w-full">
               <AppSidebar mailboxSlug={mailboxSlug} />
               <main className="flex-1 min-w-0">{children}</main>