--- a/app/(dashboard)/mailboxes/[mailbox_slug]/settings/preferences/preferencesSetting.tsx
+++ b/app/(dashboard)/mailboxes/[mailbox_slug]/settings/preferences/preferencesSetting.tsx
@@ -2,7 +2,6 @@ import { Skeleton } from "@/components/ui/skeleton";
 import { RouterOutputs } from "@/trpc";
 import ConfettiSetting from "./confettiSetting";
 import MailboxNameSetting from "./mailboxNameSetting";
-import ThemeSetting from "./themeSetting";
 
 const PreferencesSetting = ({ mailbox }: { mailbox: RouterOutputs["mailbox"]["get"] }) => {
   return (
@@ -11,11 +10,9 @@ const PreferencesSetting = ({ mailbox }: { mailbox: RouterOutputs["mailbox"]["ge
         <>
           <MailboxNameSetting mailbox={mailbox} />
           <ConfettiSetting mailbox={mailbox} />
-          <ThemeSetting mailbox={mailbox} />
         </>
       ) : (
         <>
-          <Skeleton className="h-12 w-full" />
           <Skeleton className="h-12 w-full" />
           <Skeleton className="h-12 w-full" />
         </>