--- a/app/(embed)/widget/embed/page.tsx
+++ b/app/(embed)/widget/embed/page.tsx
@@ -15,13 +15,11 @@ import PromptDetailsModal from "@/components/widget/PromptDetailsModal";
 import { useWidgetView } from "@/components/widget/useWidgetView";
 import { useScreenshotStore } from "@/components/widget/widgetState";
 import { captureExceptionAndLog } from "@/lib/shared/sentry";
-import { buildThemeCss, type MailboxTheme } from "@/lib/themes";
 import { sendConversationUpdate, sendReadyMessage } from "@/lib/widget/messages";
 import { GuideInstructions } from "@/types/guide";
 
 type DecodedPayload = {
   isWhitelabel?: boolean;
-  theme?: MailboxTheme;
   title?: string;
   exp?: number;
   iat?: number;
@@ -33,7 +31,6 @@ const GUMROAD_MAILBOX_SLUG = "gumroad";
 export default function Page() {
   const [token, setToken] = useState<string | null>(null);
   const [config, setConfig] = useState<HelperWidgetConfig | null>(null);
-  const [theme, setTheme] = useState<MailboxTheme | null>(null);
   const [defaultTitle, setDefaultTitle] = useState<string | null>(null);
   const [currentURL, setCurrentURL] = useState<string | null>(null);
   const [selectedConversationSlug, setSelectedConversationSlug] = useState<string | null>(null);
@@ -112,7 +109,6 @@ export default function Page() {
 
         try {
           const payload = jwtDecode<DecodedPayload>(content.sessionToken);
-          setTheme(payload?.theme);
           setDefaultTitle(payload?.title ?? null);
         } catch (error) {
           captureExceptionAndLog(error);
@@ -143,17 +139,6 @@ export default function Page() {
 
   return (
     <QueryClientProvider client={queryClient}>
-      <style>
-        {buildThemeCss(
-          theme ?? {
-            background: "#ffffff",
-            foreground: "#000000",
-            primary: "#000000",
-            accent: "#000000",
-            sidebarBackground: "#ffffff",
-          },
-        )}
-      </style>
       <div
         className={cx("light flex h-screen w-full flex-col responsive-chat max-w-full sm:max-w-[520px]", {
           "bg-gumroad-bg": isGumroadTheme,