--- a/components/themeProvider.tsx
+++ b/components/themeProvider.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import { ThemeProvider as NextThemesProvider, ThemeProviderProps } from "next-themes";
-
-export function ThemeProvider({ children, ...props }: ThemeProviderProps) {
-  return <NextThemesProvider {...props}>{children}</NextThemesProvider>;
-}