--- a/db/schema/mailboxes.ts
+++ b/db/schema/mailboxes.ts
@@ -35,13 +35,6 @@ export const mailboxes = pgTable(
     preferences: jsonb()
       .$type<{
         confetti?: boolean;
-        theme?: {
-          background: string;
-          foreground: string;
-          primary: string;
-          accent: string;
-          sidebarBackground: string;
-        } | null;
         autoRespondEmailToChat?: "draft" | "reply" | null;
         disableTicketResponseTimeAlerts?: boolean;
       }>()