… app/(dashboard)/mailboxes/[mailbox_slug]/settings/preferences/colorInput.tsx
… app/(dashboard)/mailboxes/[mailbox_slug]/settings/preferences/themeSetting.tsx
… docs/theming.md
… lib/themes.ts