Summary

  Implements issue #671 by removing database-driven theming and replacing with CSS-only approach. Includes
  comprehensive documentation and prepares for blue theme implementation.

  Before vs After Theme

  Before (Database-Driven Red Theme)
  
```css
  --background: hsl(0 0% 100%);           /* Plain white */
  --foreground: hsl(0 58% 10%);           /* Dark red-brown text */
  --primary: hsl(0 67% 17%);              /* Dark red buttons */
  --sidebar-background: hsl(0 69% 17%);   /* Dark red sidebar */
  --border: hsl(218 10% 84%);             /* Plain grey borders */
```
![Screenshot from 2025-07-06 02-05-53](https://github.com/user-attachments/assets/29c24058-2ef3-4482-8dbb-dbab857c2dc9)

  After (CSS-Only Blue Theme Ready)
```css
  --background: hsl(220 50% 95%);         /* Light blue background */
  --foreground: hsl(220 40% 15%);         /* Dark blue text */
  --primary: hsl(220 100% 50%);           /* Bright blue buttons */
  --sidebar-background: hsl(220 100% 15%); /* Deep blue sidebar */
  --border: hsl(220 30% 75%);             /* Blue-grey borders */
```
![Screenshot from 2025-07-06 02-04-36](https://github.com/user-attachments/assets/ab14eb3e-70f8-4d78-8c29-5fdd4cea4c9f)

  Changes Made

  🗑️ Removed

  - Database theme storage system
  - Theme UI components (ThemeProvider, themeSetting, colorInput)
  - Dynamic theme generation utilities

  ➕ Added

  - docs/theming.md - Complete theming guide (8000+ words)
  - CSS-only theming architecture
  - 6 pre-built theme examples

  📝 Modified

  - Removed theme contexts and providers
  - Simplified component architecture
  - Updated database schema

  Key Features

  - Real-time theming: Edit app/globals.css for instant changes
  - Complete documentation: Step-by-step guides and examples
  - Accessibility compliant: All color combinations tested
  - Performance improved: No runtime theme calculations

  Closes #671


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

* **New Features**
  * Added a comprehensive theming guide to the documentation, detailing how to customize the application's appearance using CSS variables.

* **Bug Fixes**
  * Improved stability by replacing the use of the native random UUID generator with a custom implementation for generating unique tab IDs.

* **Refactor**
  * Removed all custom theme management features and related UI from authentication, mailbox, and widget components.
  * Theme selection and customization options have been eliminated from user settings and preferences.
  * Simplified layout and component structures by removing theme providers and context.

* **Chores**
  * Cleaned up unused theme-related files, types, and utilities across the codebase.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->